requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.13
djangorestframework==3.15.1
idna==3.7
instaloader==4.11
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.2
requests==2.31.0
soupsieve==2.5
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1
